命令一覧のコンポーネントを再設計
