+++ demo/index.html
@@ -34,8 +34,7 @@
     <p id="info" class="info"></p>
   </div>
   <!-- command-list -->
-  <div id="command-list-button"></div>
-  <div id="command-list"></div>
+  <div id="command-list-component"></div>
 </div>
 
 <br />
